+++ changelog.md
@@ -16,9 +16,10 @@
   - Code editor:
     - Change WATERSOURCE to WATERDRINK to avoid naming issues in code complete
     - Add waterdrink doc references to pier
-  - Update mathjs 12.2.0 to 12.2.1
-  - Update electron 28.0.0 to 28.1.3
+  - Update mathjs 12.2.0 to 12.3.0
+  - Update electron 28.0.0 to 28.1.4
   - Update moment 2.29.4 to 2.30.1
+  - Update better-sqlite3 9.2.2 to 9.3.0
 
 ## 1.2.0 2023-12-18